requirements.txt
asgiref==3.7.2
Django==5.0.1
mysqlclient==2.2.1
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
